Checkstyle: Commands and Parsers (#51)
