+++ src/main/java/seedu/address/logic/commands/addcommand/AddCommand.java
@@ -1,7 +1,11 @@
 package seedu.address.logic.commands.addcommand;
 
 import seedu.address.logic.commands.Command;
+import seedu.address.model.entity.Entity;
 
+/**
+ * Adds an {@link Entity} to Alfred.
+ */
 public abstract class AddCommand extends Command {
 
     /* Possible Fields */
