+++ src/main/java/seedu/address/logic/commands/deletecommand/DeleteTeamCommand.java
@@ -6,7 +6,11 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.entity.Id;
+import seedu.address.model.entity.Team;
 
+/**
+ * Deletes a {@link Team} in Alfred.
+ */
 public class DeleteTeamCommand extends DeleteCommand {
 
     /* Possible Fields? */
