+++ src/main/java/seedu/address/logic/commands/editcommand/EditTeamCommand.java
@@ -8,6 +8,9 @@ import seedu.address.model.Model;
 import seedu.address.model.entity.Id;
 import seedu.address.model.entity.Team;
 
+/**
+ * Edits a {@link Team} in Alfred.
+ */
 public class EditTeamCommand extends EditCommand {
 
     /* Possible Fields */
@@ -28,14 +31,26 @@ public class EditTeamCommand extends EditCommand {
         return new CommandResult("");
     }
 
+    /**
+     * Creates and returns a new {@code Team} with the details {@code teamToEdit}
+     * edited with {@code editTeamDescriptor}.
+     *
+     * @param teamToEdit {@code Team} that will be updated.
+     * @param editTeamDescriptor Descriptor with the details to edit {@code teamToEdit}.
+     * @return Updated {@code Team}.
+     */
     private Team createEditedTeam(Team teamToEdit, EditTeamDescriptor editTeamDescriptor) {
         // Set each field to updated value
         // See EditCommand#EditPersonDescriptor for more context
 
-//        return new Team(/* Necessary Fields */);
+        // return new Team(/* Necessary Fields */);
         return null;
     }
 
+    /**
+     * Stores the details to edit the {@code Team} with. Each non-empty field value will replace the
+     * corresponding field value of the {@code Team}.
+     */
     public static class EditTeamDescriptor extends EditEntityDescriptor {
 
         /*
