+++ src/main/java/seedu/address/logic/commands/listcommand/ListMentorCommand.java
@@ -4,7 +4,11 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
+import seedu.address.model.entity.Mentor;
 
+/**
+ * Lists every {@link Mentor} in Alfred.
+ */
 public class ListMentorCommand extends ListCommand {
 
     /* Possible Fields? */
