+++ src/main/java/seedu/address/logic/commands/viewcommand/ViewCommand.java
@@ -3,8 +3,12 @@ package seedu.address.logic.commands.viewcommand;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.Command;
+import seedu.address.model.entity.Entity;
 import seedu.address.model.entity.Id;
 
+/**
+ * Shows detailed view of the {@link Entity} at specified ID.
+ */
 public abstract class ViewCommand extends Command {
 
     /* Possible Fields */
