+++ src/main/java/seedu/address/logic/commands/viewcommand/ViewMentorCommand.java
@@ -5,7 +5,11 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.entity.Id;
+import seedu.address.model.entity.Mentor;
 
+/**
+ * Shows detailed view of the {@link Mentor} at specified ID.
+ */
 public class ViewMentorCommand extends ViewCommand {
 
     /* Possible Fields? */
