+++ src/main/java/seedu/address/logic/commands/viewcommand/ViewTeamCommand.java
@@ -5,7 +5,11 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.entity.Id;
+import seedu.address.model.entity.Team;
 
+/**
+ * Shows detailed view of the {@link Team} at specified ID.
+ */
 public class ViewTeamCommand extends ViewCommand {
 
     /* Possible Fields? */
