+++ src/main/java/seedu/address/logic/parser/addcommandparser/AddParticipantCommandParser.java
@@ -1,13 +1,14 @@
-package seedu.address.logic.parser.AddCommandParsers;
+package seedu.address.logic.parser.addcommandparser;
 
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
+
 import seedu.address.logic.commands.addcommand.AddParticipantCommand;
+import seedu.address.logic.parser.AlfredParserUtil;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
-import seedu.address.logic.parser.AlfredParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.entity.Email;
 import seedu.address.model.entity.Id;
@@ -17,14 +18,14 @@ import seedu.address.model.entity.Phone;
 import seedu.address.model.entitylist.ParticipantList;
 
 /**
- * Parses input arguments and creates a new AddCommand object
+ * Parses input arguments and creates a new {@link AddParticipantCommand} object.
  */
 public class AddParticipantCommandParser implements Parser<AddParticipantCommand> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the AddCommand
-     * and returns an AddCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
+     * Parses the given {@code String} of arguments in the context of the {@code AddParticipantCommand}
+     * and returns an {@code AddParticipantCommand} object for execution.
+     * @throws ParseException If the user input does not conform the expected format.
      */
     public AddParticipantCommand parse(String args) throws ParseException {
         ArgumentMultimap argMultimap =
