+++ src/main/java/seedu/address/logic/parser/addcommandparser/AddTeamCommandParser.java
@@ -1,19 +1,21 @@
-package seedu.address.logic.parser.AddCommandParsers;
+package seedu.address.logic.parser.addcommandparser;
 
 import static seedu.address.logic.parser.CliSyntax.PREFIX_LOCATION;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PROJECT_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PROJECT_TYPE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_SUBJECT_NAME;
+
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
+
 import seedu.address.logic.commands.addcommand.AddCommand;
 import seedu.address.logic.commands.addcommand.AddTeamCommand;
+import seedu.address.logic.parser.AlfredParserUtil;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
-import seedu.address.logic.parser.AlfredParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.entity.Id;
 import seedu.address.model.entity.Location;
@@ -27,7 +29,7 @@ import seedu.address.model.entity.Team;
 import seedu.address.model.entitylist.TeamList;
 
 /**
- * Parses input arguments and creates a new AddCommand object
+ * Parses input arguments and creates a new {@link AddTeamCommand} object.
  */
 public class AddTeamCommandParser implements Parser<AddCommand> {
