+++ src/main/java/seedu/address/logic/parser/deletecommandparser/DeleteMentorCommandParser.java
@@ -1,21 +1,27 @@
-package seedu.address.logic.parser.DeleteCommandParsers;
+package seedu.address.logic.parser.deletecommandparser;
+
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.logic.commands.deletecommand.DeleteCommand;
 import seedu.address.logic.commands.deletecommand.DeleteMentorCommand;
 import seedu.address.logic.parser.AlfredParserUtil;
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.entity.Id;
 import seedu.address.model.entity.PrefixType;
 
 /**
- * Parses input arguments and creates a new DeleteCommand object
+ * Parses input arguments and creates a new {@link DeleteMentorCommand} object.
  */
 public class DeleteMentorCommandParser implements Parser<DeleteCommand> {
 
     private static final String MESSAGE_USAGE = "Wrong usage.";
 
+    /**
+     * Parses the given {@code String} of arguments in the context of the {@code DeleteMentorCommand}
+     * and returns a {@code DeleteMentorCommand} object for execution.
+     * @throws ParseException if the user input does not conform the expected format
+     */
     @Override
     public DeleteMentorCommand parse(String args) throws ParseException {
         try {
