+++ src/main/java/seedu/address/logic/parser/deletecommandparser/DeleteParticipantCommandParser.java
@@ -1,6 +1,7 @@
-package seedu.address.logic.parser.DeleteCommandParsers;
+package seedu.address.logic.parser.deletecommandparser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
 import seedu.address.logic.commands.deletecommand.DeleteCommand;
 import seedu.address.logic.commands.deletecommand.DeleteParticipantCommand;
 import seedu.address.logic.parser.AlfredParserUtil;
@@ -10,15 +11,15 @@ import seedu.address.model.entity.Id;
 import seedu.address.model.entity.PrefixType;
 
 /**
- * Parses input arguments and creates a new DeleteCommand object
+ * Parses input arguments and creates a new {@link DeleteParticipantCommand} object.
  */
 public class DeleteParticipantCommandParser implements Parser<DeleteCommand> {
 
     private static final String MESSAGE_USAGE = "Wrong usage.";
 
     /**
-     * Parses the given {@code String} of arguments in the context of the DeleteCommand
-     * and returns a DeleteCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the {@code DeleteParticipantCommand}
+     * and returns a {@code DeleteParticipantCommand} object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
     @Override
