+++ src/main/java/seedu/address/logic/parser/editcommandparser/EditMentorCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser.EditCommandParsers;
+package seedu.address.logic.parser.editcommandparser;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EMAIL;
@@ -6,6 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ORGANISATION;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_SUBJECT_NAME;
+
 import seedu.address.logic.commands.editcommand.EditCommand;
 import seedu.address.logic.commands.editcommand.EditMentorCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
@@ -15,14 +16,14 @@ import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.entity.Id;
 
 /**
- * Parses input arguments and creates a new EditCommand object
+ * Parses input arguments and creates a new {@link EditMentorCommand} object.
  */
 public class EditMentorCommandParser implements Parser<EditCommand> {
 
     /**
-     * Parses the given {@code String} of arguments in the context of the EditCommand
-     * and returns an EditCommand object for execution.
-     * @throws ParseException if the user input does not conform the expected format
+     * Parses the given {@code String} of arguments in the context of the {@code EditMentorCommand}
+     * and returns an {@code EditMentorCommand} object for execution.
+     * @throws ParseException if the user input does not conform the expected format.
      */
     public EditMentorCommand parse(String args) throws ParseException {
         requireNonNull(args);
