+++ src/main/java/seedu/address/logic/parser/editcommandparser/EditParticipantCommandParser.java
@@ -1,0 +1,27 @@
+package seedu.address.logic.parser.editcommandparser;
+
+import seedu.address.logic.commands.editcommand.EditCommand;
+import seedu.address.logic.commands.editcommand.EditParticipantCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+/**
+ * Parses input arguments and creates a new {@link EditParticipantCommand} object.
+ */
+public class EditParticipantCommandParser implements Parser<EditCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of the {@code EditParticipantCommand}
+     * and returns an {@code EditParticipantCommand} object for execution.
+     * @throws ParseException if the user input does not conform the expected format.
+     */
+    public EditParticipantCommand parse(String args) throws ParseException {
+
+        /*
+         * This is just placeholder code. We will implement proper code
+         * when the Participant class has been finalised.
+         */
+
+        return null;
+    }
+}
