+++ src/main/java/seedu/address/logic/parser/editcommandparser/EditTeamCommandParser.java
@@ -1,0 +1,28 @@
+package seedu.address.logic.parser.editcommandparser;
+
+import seedu.address.logic.commands.editcommand.EditCommand;
+import seedu.address.logic.commands.editcommand.EditTeamCommand;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.exceptions.ParseException;
+
+/**
+ * Parses input arguments and creates a new {@link EditTeamCommand} object.
+ */
+public class EditTeamCommandParser implements Parser<EditCommand> {
+
+    /**
+     * Parses the given {@code String} of arguments in the context of the {@code EditTeamCommand}
+     * and returns an {@code EditTeamCommand} object for execution.
+     * @throws ParseException if the user input does not conform the expected format.
+     */
+    public EditCommand parse(String args) throws ParseException {
+
+        /**
+         * This is just placeholder code. We will implement proper code
+         * when the Team class has been finalised.
+         */
+
+        return null;
+    }
+
+}
