+++ src/main/java/seedu/address/logic/parser/viewcommandparser/ViewMentorCommandParser.java
@@ -1,12 +1,13 @@
-package seedu.address.logic.parser.ViewCommandParsers;
+package seedu.address.logic.parser.viewcommandparser;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
+
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.AlfredParserUtil;
+import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
