+++ src/main/java/seedu/address/logic/parser/viewcommandparser/ViewParticipantCommandParser.java
@@ -1,19 +1,20 @@
-package seedu.address.logic.parser.ViewCommandParsers;
+package seedu.address.logic.parser.viewcommandparser;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
+
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.AlfredParserUtil;
+import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new EditCommand object
  */
-public class ViewTeamCommandParser implements Parser<EditCommand> {
+public class ViewParticipantCommandParser implements Parser<EditCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the EditCommand
