+++ src/main/java/seedu/address/logic/parser/viewcommandparser/ViewTeamCommandParser.java
@@ -1,19 +1,20 @@
-package seedu.address.logic.parser.ListCommandParsers;
+package seedu.address.logic.parser.viewcommandparser;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
+
 import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.AlfredParserUtil;
+import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
 
 /**
  * Parses input arguments and creates a new EditCommand object
  */
-public class ListMentorCommandParser implements Parser<EditCommand> {
+public class ViewTeamCommandParser implements Parser<EditCommand> {
 
     /**
      * Parses the given {@code String} of arguments in the context of the EditCommand
@@ -23,12 +24,12 @@ public class ListMentorCommandParser implements Parser<EditCommand> {
     public EditCommand parse(String args) throws ParseException {
 
         // The return type of this command will also have to be corrected
-        // to return a ListMentorCommand object as designed by John instead of an
+        // to return a ViewTeamCommand object as designed by John instead of an
         // EditCommand Object.
 
         /**
          * This is just placeholder code. We will implement proper code
-         * when the Mentor class has been finalised.
+         * when the Team class has been finalised.
          */
 
         return null;
